+++ contracts/interfaces/IyVaultV2.sol
@@ -4,44 +4,104 @@ import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IyVaultV2 is IERC20 {
     function token() external view returns (address);
-    function deposit() external returns (uint);
-    function deposit(uint) external returns (uint);
-    function deposit(uint, address) external returns (uint);
-    function withdraw() external returns (uint);
-    function withdraw(uint) external returns (uint);
-    function withdraw(uint, address) external returns (uint);
-    function withdraw(uint, address, uint) external returns (uint);
-    function permit(address, address, uint, uint, bytes32) external view returns (bool);
-    function pricePerShare() external view returns (uint);
-    
+
+    function deposit() external returns (uint256);
+
+    function deposit(uint256) external returns (uint256);
+
+    function deposit(uint256, address) external returns (uint256);
+
+    function withdraw() external returns (uint256);
+
+    function withdraw(uint256) external returns (uint256);
+
+    function withdraw(uint256, address) external returns (uint256);
+
+    function withdraw(
+        uint256,
+        address,
+        uint256
+    ) external returns (uint256);
+
+    function permit(
+        address,
+        address,
+        uint256,
+        uint256,
+        bytes32
+    ) external view returns (bool);
+
+    function pricePerShare() external view returns (uint256);
+
     function apiVersion() external view returns (string memory);
-    function totalAssets() external view returns (uint);
-    function maxAvailableShares() external view returns (uint);
-    function debtOutstanding() external view returns (uint);
-    function debtOutstanding(address strategy) external view returns (uint);
-    function creditAvailable() external view returns (uint);
-    function creditAvailable(address strategy) external view returns (uint);
-    function availableDepositLimit() external view returns (uint);
-    function expectedReturn() external view returns (uint);
-    function expectedReturn(address strategy) external view returns (uint);
+
+    function totalAssets() external view returns (uint256);
+
+    function maxAvailableShares() external view returns (uint256);
+
+    function debtOutstanding() external view returns (uint256);
+
+    function debtOutstanding(address strategy) external view returns (uint256);
+
+    function creditAvailable() external view returns (uint256);
+
+    function creditAvailable(address strategy) external view returns (uint256);
+
+    function availableDepositLimit() external view returns (uint256);
+
+    function expectedReturn() external view returns (uint256);
+
+    function expectedReturn(address strategy) external view returns (uint256);
+
     function name() external view returns (string memory);
+
     function symbol() external view returns (string memory);
-    function decimals() external view returns (uint);
-    function balanceOf(address owner) external view override returns (uint);
-    function totalSupply() external view override returns (uint);
+
+    function decimals() external view returns (uint256);
+
+    function balanceOf(address owner) external view override returns (uint256);
+
+    function totalSupply() external view override returns (uint256);
+
     function governance() external view returns (address);
+
     function management() external view returns (address);
+
     function guardian() external view returns (address);
+
     function guestList() external view returns (address);
-    function strategies(address) external view returns (uint, uint, uint, uint, uint, uint, uint, uint);
-    function withdrawalQueue(uint) external view returns (address);
+
+    function strategies(address)
+        external
+        view
+        returns (
+            uint256,
+            uint256,
+            uint256,
+            uint256,
+            uint256,
+            uint256,
+            uint256,
+            uint256
+        );
+
+    function withdrawalQueue(uint256) external view returns (address);
+
     function emergencyShutdown() external view returns (bool);
-    function depositLimit() external view returns (uint);
-    function debtRatio() external view returns (uint);
-    function totalDebt() external view returns (uint);
-    function lastReport() external view returns (uint);
-    function activation() external view returns (uint);
+
+    function depositLimit() external view returns (uint256);
+
+    function debtRatio() external view returns (uint256);
+
+    function totalDebt() external view returns (uint256);
+
+    function lastReport() external view returns (uint256);
+
+    function activation() external view returns (uint256);
+
     function rewards() external view returns (address);
-    function managementFee() external view returns (uint);
-    function performanceFee() external view returns (uint);
+
+    function managementFee() external view returns (uint256);
+
+    function performanceFee() external view returns (uint256);
 }
